moved data operations to separate file
